feat(tools): implement the remote studio view
