+++ dsdl/objectio/__init__.py
@@ -2,6 +2,7 @@ from .local import LocalFileReader
 from .ali_oss import AliOSSFileReader
 from .base import BaseFileReader
 from .ceph import CephFileReader, PetrelFileReader
+from .aws_oss import AwsOSSFileReader
 
 __all__ = [
     "LocalFileReader",
@@ -9,4 +10,5 @@ __all__ = [
     "BaseFileReader",
     "CephFileReader",
     "PetrelFileReader",
+    "AwsOSSFileReader",
 ]
