+++ dsdl/tools/__init__.py
@@ -1,9 +1,9 @@
-from .visualize import view, studio_view, StudioView
+from .visualize import view
 from .check import check
+from .studio_view import StudioView
 
 __all__ = [
     "view",
-    "studio_view",
     "check",
     "StudioView",
 ]
